refactor: split resolver for public and private parts to hide inner logic with lenient mode in future
