+++ kv/internal/resolve/helpers_test.go
@@ -1,4 +1,4 @@
-package resolver
+package resolve
 
 import (
 	"errors"
